Moved unit_vec and polygon functions to lib/utils
