+++ example_problems/electronic_boltzmann/pdcoo2/L_1.0_1.25_tau_ee_inf_tau_eph_inf_DC/README.txt
@@ -9,4 +9,8 @@ to be made :
     - Change p_dim (if switching between 1D and 2D representation in p-space)
 3. initialize : No change
 4. main : No change
-5. boundary_conditions : No change
+i5. boundary_conditions : No change
+
+
+
+PdCoO2 works only with the polar2D grid in momentum space for the moment.
